move components to components folder
